Made pickup logic
Added equipment pickups saved in prefabs when you walk up to one you can press the e key to interact and it will add the equipment to your inventory. After its in your inventory you can click and drag the icon to the character slot to equip the part. Cant get the flames to show up like it does in josh scene but I think it is equipping but could be wrong!